mcp_server/requirements.txt
fastmcp>=0.1.0
httpx>=0.24.0
python-dotenv>=0.21.0
pydantic>=2.0.0
uvicorn>=0.22.0